refactor(services): ProcessStateDisplay SDK->CLI
Move it to the command as it's concerned with user facing display. We
have to make `ProcessState` public as a result. Added some `IntoIter`
and `From` to make `ProcessStates` easier to use from the outside.
