+++ cli/flox-rust-sdk/src/providers/services.rs
@@ -5,16 +5,13 @@
 //! Note that `process-compose` terminates when all services are stopped. To prevent this, we inject
 //! a dummy service (`flox_never_exit`) that sleeps indefinitely.
 
-use std::cmp::max;
 use std::collections::BTreeMap;
 use std::env;
-use std::fmt::Display;
 use std::io::{BufRead, Read};
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 use std::sync::mpsc::{Receiver, Sender};
 
-use itertools::Itertools;
 use once_cell::sync::Lazy;
 #[cfg(test)]
 use proptest::prelude::*;
@@ -210,16 +207,16 @@ pub fn maybe_make_service_config_file(
 
 /// The parsed output of `process-compose process list` for a single process.
 #[derive(Deserialize, Debug, Clone, PartialEq)]
-struct ProcessState {
-    name: String,
+pub struct ProcessState {
+    pub name: String,
     namespace: String,
-    status: String,
+    pub status: String,
     system_time: String,
     age: u64,
     is_ready: String,
     restarts: u64,
     exit_code: i32,
-    pid: u64,
+    pub pid: u64,
     #[serde(skip_serializing, rename = "IsRunning")]
     is_running: bool,
 }
@@ -272,60 +269,18 @@ impl ProcessStates {
     }
 }
 
-#[derive(Clone, Debug, Serialize)]
-/// Simplified version of ProcessState for display in the CLI.
-pub struct ProcessStateDisplay {
-    name: String,
-    status: String,
-    pid: u64,
-}
-
-#[derive(Clone, Debug, Serialize)]
-/// Simplified version of ProcessStates for display in the CLI.
-pub struct ProcessStatesDisplay(Vec<ProcessStateDisplay>);
-
-impl From<ProcessStates> for ProcessStatesDisplay {
-    fn from(procs: ProcessStates) -> Self {
-        ProcessStatesDisplay(
-            procs
-                .0
-                .into_iter()
-                .sorted_by_key(|proc| proc.name.clone())
-                .map(|proc| ProcessStateDisplay {
-                    name: proc.name,
-                    status: proc.status,
-                    pid: proc.pid,
-                })
-                .collect(),
-        )
-    }
-}
-
-impl IntoIterator for ProcessStatesDisplay {
-    type IntoIter = std::vec::IntoIter<ProcessStateDisplay>;
-    type Item = ProcessStateDisplay;
+impl IntoIterator for ProcessStates {
+    type IntoIter = std::vec::IntoIter<ProcessState>;
+    type Item = ProcessState;
 
     fn into_iter(self) -> Self::IntoIter {
         self.0.into_iter()
     }
 }
 
-impl Display for ProcessStatesDisplay {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let name_width_min = 10;
-        let name_width = max(
-            name_width_min,
-            self.0.iter().map(|proc| proc.name.len()).max().unwrap_or(0),
-        );
-        writeln!(f, "{:<name_width$} {:<10} {:>8}", "NAME", "STATUS", "PID")?;
-        for proc in &self.0 {
-            writeln!(
-                f,
-                "{:<name_width$} {:<10} {:>8}",
-                proc.name, proc.status, proc.pid
-            )?;
-        }
-        Ok(())
+impl From<Vec<ProcessState>> for ProcessStates {
+    fn from(procs: Vec<ProcessState>) -> Self {
+        ProcessStates(procs)
     }
 }
 
@@ -608,6 +563,26 @@ impl ProcessComposeLogReader {
     }
 }
 
+pub mod test_helpers {
+    use super::*;
+
+    /// Shorthand for generating a ProcessState with fields that we care about.
+    pub fn generate_process_state(name: &str, status: &str, pid: u64) -> ProcessState {
+        ProcessState {
+            name: name.to_string(),
+            namespace: "".to_string(),
+            status: status.to_string(),
+            system_time: "".to_string(),
+            age: 0,
+            is_ready: "".to_string(),
+            restarts: 0,
+            exit_code: 0,
+            pid,
+            is_running: true,
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
@@ -852,97 +827,5 @@ mod tests {
             "expected no more messages, got: {:?}",
             remaining_messages
         );
-    }
-
-    /// Shorthand for generating a ProcessState with fields that we care about.
-    fn generate_process_state(name: &str, status: &str, pid: u64) -> ProcessState {
-        ProcessState {
-            name: name.to_string(),
-            namespace: "".to_string(),
-            status: status.to_string(),
-            system_time: "".to_string(),
-            age: 0,
-            is_ready: "".to_string(),
-            restarts: 0,
-            exit_code: 0,
-            pid,
-            is_running: true,
-        }
-    }
-
-    #[test]
-    fn test_processstatesdisplay_name_sorted() {
-        let states = ProcessStates {
-            0: vec![
-                generate_process_state("bbb", "Running", 123),
-                generate_process_state("zzz", "Running", 123),
-                generate_process_state("aaa", "Running", 123),
-                generate_process_state("ccc", "Running", 123),
-            ],
-        };
-        let states_display: ProcessStatesDisplay = states.into();
-        assert_eq!(format!("{states_display}"), indoc! {"
-            NAME       STATUS          PID
-            aaa        Running         123
-            bbb        Running         123
-            ccc        Running         123
-            zzz        Running         123
-        "});
-    }
-
-    #[test]
-    fn test_processstatesdisplay_name_padded() {
-        let states = ProcessStates {
-            0: vec![
-                generate_process_state("short", "Running", 123),
-                generate_process_state("longlonglonglonglong", "Running", 123),
-            ],
-        };
-        let states_display: ProcessStatesDisplay = states.into();
-        assert_eq!(format!("{states_display}"), indoc! {"
-            NAME                 STATUS          PID
-            longlonglonglonglong Running         123
-            short                Running         123
-        "});
-    }
-
-    #[test]
-    fn test_processstatesdisplay_status_variants() {
-        let states = ProcessStates {
-            0: vec![
-                generate_process_state("aaa", "Running", 123),
-                generate_process_state("bbb", "Stopped", 123),
-                generate_process_state("ccc", "Completed", 123),
-            ],
-        };
-        let states_display: ProcessStatesDisplay = states.into();
-        assert_eq!(format!("{states_display}"), indoc! {"
-            NAME       STATUS          PID
-            aaa        Running         123
-            bbb        Stopped         123
-            ccc        Completed       123
-        "});
-    }
-
-    #[test]
-    fn test_processstatesdisplay_pid_aligned() {
-        let states = ProcessStates {
-            0: vec![
-                generate_process_state("aaa", "Running", 1),
-                generate_process_state("bbb", "Running", 12),
-                generate_process_state("ccc", "Running", 123),
-                generate_process_state("ddd", "Running", 1234),
-                generate_process_state("eee", "Running", 12345),
-            ],
-        };
-        let states_display: ProcessStatesDisplay = states.into();
-        assert_eq!(format!("{states_display}"), indoc! {"
-            NAME       STATUS          PID
-            aaa        Running           1
-            bbb        Running          12
-            ccc        Running         123
-            ddd        Running        1234
-            eee        Running       12345
-        "});
     }
 }
